requirements.txt
clip==0.2.0
cog==0.4.4
japanese_clip==0.2.0
MeCab==0.996.3
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.2.0
pycocoevalcap==1.2
requests==2.28.1
skimage==0.0
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.64.0
transformers==4.22.2